fix bug users

+++ database/migrations/2024_07_15_122631_add_role_to_users_table.php
@@ -9,17 +9,24 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-        public function up()
+    public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->enum('role', ['admin', 'staff', 'customer'])->default('customer');
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->enum('role', ['admin', 'staff', 'customer'])->default('customer');
+            }
         });
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('role');
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
